Refactor UserAuthFetcher and move the helper structs in auth to authutil
Favour value instead of pointer in UserAuthFetcher and put the input and
output variable to function parameters instead of struct properties.

Make buildAuthDataQuery construct the predicate in a simpler way.
